+++ feincms3/plugins/external.py
@@ -109,10 +109,10 @@ class External(models.Model):
     alternative_text = models.CharField(
         _("alternative text"),
         help_text=_("Describe the contents, e.g. for screenreaders."),
-        max_length=200,
+        max_length=1000,
         blank=True,
     )
-    caption = models.CharField(_("caption"), blank=True, max_length=200)
+    caption = models.CharField(_("caption"), blank=True, max_length=1000)
 
     class Meta:
         abstract = True